Re-implement the Tokenizer class from scratch
It didn't have a license header. Even though we received explicit authorization
to use it, it wasn't a comfortable situation.

This is an original implementation that only keeps the APIs we use in the
parameter-framework and keep them identical (except for a small exception). The
behaviour is also exactly the same.

Change-Id: I85a69c76027ee026a693d79cd19edd3b86796f9d
Signed-off-by: David Wagner <david.wagner@intel.com>