[coverage] Correct line number msg
When an error occurs during input parsing,
the line number was incorrectly displayed.

In python, enumerating an iterable will yield
an pair of index and a value.
This index starts at 0: next(enumerate([""])) == (0, "")

This index was displayed as the line number.

Add one to the index to get the line number.

Signed-off-by: Kevin Rocard <kevin.rocard@intel.com>